Trying to get .env to work AND blocking feature to be enabled in the cargo.toml. Currently I have to manually call --features='reqwest-blocking'
