+++ client/.env
@@ -1,0 +1,3 @@
+AUTH0_CLIENT_ID="BnJiVrOXavZ1mbvsiwvBcZ96dTFH9k4L"
+
+AUTH0_ISSUER_URL="https://dev-k6zdm2f3z3kst6r7.us.auth0.com"
