+++ client/Cargo.toml
@@ -6,18 +6,26 @@ authors = ["Karl Nyborg"]
 rust-version = "1.86.0"
 
 [dependencies]
+anyhow = "1.0.98"
+dotenv = "0.15.0"
 egui = "0.21.0"
 egui-macroquad = "0.15.0"
 egui_demo_lib = "0.21.0"
+env_logger = "0.11.8"
 macroquad = "0.3.26"
+oauth2 = "5.0.0"
 once_cell = "1.21.3"
 openidconnect = "4.0.0"
 rayon = { version = "1.10.0", optional = true }
+reqwest = { version = "0.12.15", features = ["blocking", "rustls-tls"], default-features = false }
 secs = { git = "https://github.com/wick3dr0se/secs", version = "0.1.0" }
 spacetimedb-sdk = "1.1.1"
+url = "2.5.4"
 
 [features]
 multithreaded = ["rayon"]
+reqwest = ["oauth2/reqwest"]
+reqwest-blocking = ["oauth2/reqwest-blocking"]
 
 [package.metadata.rust-analyzer]
 # Enable experimental features in rust-analyzer
